Move evolution-tree without changes to `shared`
